Adds initial code to network part WiP
Work in progress but this change adds client and server side of app.
Idea is that application shall run with flag --server if it is inteded
to run server and without any flag (or with flag --client) to run
client.

I moved "Hello world" code from main to client.cpp because only client
will have GUI.